Improve tracking of upstream changes
This patch removes the "MODIFICATION" tags from the layout files and places .upstream files instead.
Similarly, the existing .original template files are renamed to .upstream files.

This is done to unclutter the Boost Union files and prepare them for more extensive changes, but to still be able to track changes in upstream Boost Core and detect necessary backports to Boost Union.