Add course related hints feature, solves #5
This patch ports the 'course related hints' feature from Boost Campus to Boost Union.

With this patch, four new settings are introduced:
* Position of switch role information
* Show hint in hidden courses
* Show hint guest for access
* Show hint for self enrolment without enrolment key

The implementation is more or less the same as it was in Boost Campus, but the rendering of the hints are changed from html_writer to Mustache templates.