Feature: Back to top button, solves #7
This patch adds an admin setting which allows the admin to add a back to top button to the page. The button will appear in the bottom right corner of the page as soon as the user scrolls down the page.

A button like this existed already on Boost in Moodle Core until Moodle 3.11, but was removed in 4.0. With Boost Union, you can bring it back.