Feature: Built-in imprint, solves #32
This patch allows the admin to enable a imprint page which can be filled with custom content.
Additionally, the link to the imprint page can be added to the footnote or the footer automatically.

This feature is particularly intended to be an easier-to-use replacement for local_staticpage and other third party solutions if you just need to add an imprint to your site.