Feature: Random login background image with text, solves #36.
With this patch, the admin can not only add one but up to 25 files as a background image for the login page. One of these images will be picked randomly and shown when the user visits the login page.
Additionally, the admin can add text, e.g. a copyright notice to the uploaded login background images.