Add new login layouts and other login settings resolve 1055
This commit adds two new loginlayouts (tabs and accordion) aswell as a loginbackgroundlayout option to switch between the default layout and a splitscreen layout.
Beyond that this commit adds options to control other aspects of the login page for example the logincontainerwidth, the logintexts and the option to disable the individual loginmethods.